Kill 'CourseEntryNTIID'.
First, clients are not and must not be aware of any hierarchy. This is purely an internal implementation detail.

Second and more importantly, in most circumstances where a client has access to see a section catalog entry, they simply will not be able to see the main course catalog entry in the first place. That fact is hidden from them, so this value was meaningless and wouldn't correlate with anything.



git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.products.courseware/trunk@45461 3dfb626b-079f-4ab1-a9f6-4654e5377508